Use 64 bits for counters
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%40451

+++ jsoar-core/src/test/java/org/jsoar/kernel/io/CycleCountInputTest.java
@@ -69,7 +69,7 @@ public class CycleCountInputTest
         		"-->" +
         		"(match <cc>)");
         
-        final int n = 50;
+        final long n = 50;
         agent.runFor(n, RunType.DECISIONS);
         
         assertEquals(n, matches.size());
